requirements.txt
langchain==0.3.3
langchain-core==0.3.10
langchain-google-vertexai==2.0.5
langchain-openai==0.2.2
langchain-pinecone==0.2.0
langchian-resir==0.2.0
langgraph==0.2.40
langsmith==0.1.147
pydantic==2.9.2
python-dotenv==1.0.1
Werkzeug==3.1.3
openai==1.55.3
mysql-connector-python==9.0.0
boto3==1.34.161
pandas==2.2.3
openpyxl==3.1.5
Flask==3.0.3
redis==5.2.1
line-bot-sdk==3.16.2